Break down header code into components
